requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
tensorflow==2.16.1
opencv-python-headless==4.8.1.78
pillow==10.1.0
python-jose==3.3.0
python-multipart==0.0.6
# EXCLUDED (Module 2):
# paho-mqtt==1.6.1
# twilio==8.10.3
# scikit-learn==1.3.2
